image_restoration/requirements.txt
numpy
pillow
opencv-python-headless
streamlit
streamlit-drawable-canvas